requirements.txt
# fastapi==0.104.1
# uvicorn==0.24.0
# kafka-python==2.0.2
# redis==5.0.1
# pandas==2.0.3
# numpy==1.24.3
# scikit-learn==1.3.0
# pydantic==2.4.2
# requests==2.31.0

fastapi
uvicorn
kafka-python
redis
pandas
numpy
scikit-learn
pydantic
requests